fixed some misc problems
