+++ Final-Project/server/models/users.js
@@ -80,17 +80,17 @@ async function remove(id) {
     return null;
 }
 
-async function login(handle) {
-    const data = await dataP;
-    const user = data.find(item => item.handle === handle);
+// async function login(handle) {
+//     const data = await dataP;
+//     const user = data.find(item => item.handle === handle);
 
-    if (!user) {
-        throw new Error("Invalid user handle");
-    }
+//     if (!user) {
+//         throw new Error("Invalid user handle");
+//     }
 
-    return user;
-}
+//     return user;
+// }
 
 module.exports = {
-    getAll, get, search, add, update, remove, login
+    getAll, get, search, add, update, remove
 }
